+++ CHANGELOG.md
@@ -3,5 +3,11 @@ Only partial changelog, [commit history](https://framagit.org/SondagePro-LimeSur
 
 ## [unreleased]
 
+### Feature
+
+- Save in Upload file question type
+- Allow replace print answers
+- Allow to send to any email
+
 ### Dev
 - Start with [sendPdfReport release 1.2.2](https://framagit.org/SondagePro-LimeSurvey-plugin/sendPdfReport/tags/1.2.2)